Убрал main.storyboard
