+++ FibExampleApp/FibExampleApp/AppDelegate.swift
@@ -14,35 +14,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 	func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
 		// Override point for customization after application launch.
-//		let view = UIView()
-//		view.backgroundColor = .black
-		FibViewController.defaultConfiguration = .init(viewConfiguration: .init(roundedShutterBackground: .black.withAlphaComponent(0.2), shutterBackground: .black.withAlphaComponent(0.4), viewBackgroundColor: .white, shutterType: .rounded, backgroundView: nil))
-		RoundedCell.defaultRoundedCellAppearance.shadowClosure = { view in
-			view.layer.shadowColor = UIColor.black.cgColor
-			view.layer.shadowRadius = 10
-			view.layer.shadowOpacity = 1
-		}
-		let appearance = UINavigationBarAppearance()
-		appearance.configureWithTransparentBackground()
-		appearance.backgroundColor = UIColor.clear
-		appearance.backgroundEffect = UIBlurEffect(style: .dark) // or dark
-		appearance.titleTextAttributes = [.foregroundColor: UIColor.white]
-		
-		let scrollingAppearance = UINavigationBarAppearance()
-		scrollingAppearance.configureWithTransparentBackground()
-//		scrollingAppearance.backgroundColor = .white // your view (superview) color
-		
-		UINavigationBar.appearance().standardAppearance = appearance
-		UINavigationBar.appearance().scrollEdgeAppearance = appearance
-		UINavigationBar.appearance().compactAppearance = appearance
-		DispatchQueue.main.async {
-			let vc = ViewController()
-			let nav = UINavigationController(rootViewController: vc)
-			let window = UIWindow(frame: UIScreen.main.bounds)
-			window.rootViewController = nav
-			nav.view.backgroundColor = .alizarin
-			window.makeKeyAndVisible()
-		}
 		return true
 	}
